+++ packaging/nix/build-nix-flake.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+# Resolve script directory and change to it
+cd "$(dirname "$0")"
+
+echo "Building Nix derivation package..."
+nix-build default.nix